requirements.txt
pygame>=2.5
sounddevice>=0.4
soundfile>=0.12
PyYAML>=6.0
numpy>=1.24
